3º ANO/Projeto-SD/requirements.txt
# Web Framework
fastapi
uvicorn[standard]
websockets
jinja2
python-multipart

# HTTP Client
requests

# gRPC
grpcio
grpcio-tools
protobuf

# Utilitários
python-dotenv
pybloom_live
cachetools

# Web Scraping
beautifulsoup4
lxml

# IA
google-generativeai
groq
httpx